Regigged the dummy monitoring screen. Made the layout more compact, moved the learner functionality to the learner screens.
Fixed a transaction bug with startLesson - the method was not named consistently in the applicationContext file, so the method was not part of a proper transaction. As such, the db changes done by the method were lost. This resulted in the lesson staying in the created state, never going onto the started state. Made the naming consistent and it now goes into the started state as expected.