LDEV-73: Added a method to pass the details of the license table to Flash. The DTO code in lams.jar tweaks the data before sending - in this case we make sure the picture url is a complete url (maybe relative to the current server in db).
LDEV-77: Issues exist with the data returned by getAccessibleWorkspaceFolders(). Data is missing, and the structure is different to the workspace data returned by getFolderContentsExcludeHome().

Trying to fix the problems by changing the DTO to be the same one used for getFolderContentsExcludeHome(). Put it in a temporary method for the moment so that we can try out the changed packet without breaking the existing code. More changes will probably be required for this problem.