Made changes as per wiki:
In Eclipse text files default to "-kk". To change the mode of the files, click on a file or a folder and select "Team -> Change ASCII/Binary Property" from the popup menu. Change text files to "ASCII with keyword expansion (-kkv)". It is not possible to change binary files to support keyword expansion.