New feature from Ernie: Shared Resources Authoring
  Tool name in the left panel is "Share Resoruces". It should be "Share Resources"
  Drop down menu for the resources available in Basic


Monitor

even when no student has reached the activity the "Edit Activity" tab says "This Activity is no longer editable"

In Summary: We need to add the time and date that the student has access a resource

In Summary: Can we hide/show the entries from the teacher as well ?

In Summary: can the title column have the hyperlink to the resource?
(just in case the teacher wants to view it)

Instructions: view the instructions and files.

Learner

The name of the person that has added a resource is not being displayed (when students can add resources/URLs)