Modifying the getOrganisationForUserByRole calls to return just the course and class when specified, rather than the whole organisation tree.
Replaced getActiveLessons  with getLesson, to suit the change from listing all the lessons in the learner client to just working with a single lesson.

Moved the logic to cache the lesson learners from learning into the lams.jar (in the LessonService).

Replaced getAllLessonsWDDX, getLessonData  with getLessonDetails. Suits  the change from listing all the lessons in the monitoring client to just working with a single lesson.