Random grouping now working.
Fixed issues in authoring - was trying to look up grouping by UUID and that won't work for more than one design with grouping (as the UUIDs may be duplicated). Modified authoring to get all the existing groupings via the grouping activities.

Changed to make learner service changes to stop a hibernate lazy initialisation exception when grouping has just been run. Also reduced the amount of data needed to be cached in the session - only the lesson id is needed not the complete lesson.