LDEV-1676: Adding a new column in forun that shows if any of a learner's files have been marked.
Also fixed a problem with the view forum screen in teacher mode. When accessed from monitoring, the teacher was able to see and edit the learner's reflection. Now when the screen is accessed from the summary screen, the reflection cannot be seen (as this is primarily for marking). When the screen is accessed from the learner progress screen the reflection can be seen but not modified.