LDEV-1657 - Updating the wiki link plugin for the fckeditor.
To attempt to resolve issues with certain browsers, I removed any ajax calls to get the list of wiki links. Instead, they are retrieved from the opening page. ie a javascript array of wikilinks is created on the page where the fckeditor is, then it is retrieved by the plugin with the following command:

wikiArray = window.top.opener.top.document.getElementById("wikiLinks").value.split(',');