LDEV-2577 The new CKEditor does not have Java Integration library (yet). We have been using it to: a) get the tag library to create our own tag <lams:FCKeditor> But the new CKEditor uses AJAX to get the contents, so the library is not needed anymore. The references to it has been deleted from every tool. FCKEditor.tag has been rewritten to use the new syntax. b) get a FCKEditor internal file manager by a Connector from the jar The new CKEditor does not provide a file manager - the license became commercial and the whole thing became a separate project, CKFinder. There were alternatives suggested on the web: jasfinder - http://sourceforge.net/projects/jasfinder/support kcfinder - http://kcfinder.sunhater.com/ some other - http://www.miguelmanchego.com/2009/gestor-de-archivos-ckeditor/ but we have already nicely integrated with the old file manager. This page: http://jacek-kromski.jogger.pl/2009/09/25/integracja-ckeditor-z-przegladarka-plikow-z-fckeditor/ in polish says how to use the file manager from the old FCKEditor. The steps that have to be done: - download the new CKEditor - download the old FCKEditor - copy the filemanager to ckeditor directory - in the file browser/default/frmresourceslist.html, line 92 (with SetUrl call) needs to be replaced to: window.top.opener.CKEDITOR.tools.callFunction( window.top.document.CKEditorFuncNum, fileUrl, '' ); - in the file browser/default/browser.html add in line 80: document.CKEditorFuncNum = GetUrlParam( 'CKEditorFuncNum');
FCKeditor has been used in version 2.6.6, the last one before switching to CKEditor project.
Changes required to use newer version of Java Integration library
for FCKEditor so it was upgraded to the latest (for FCKEditor, because one for CKEditor does not exist) version 2.6.

This also required the Prototype to be upgraded to the latest 1.6.1.

The CKEditor code also needed  to be changed, so if  another upgrade to a newer version occurs, we need to remember to repeat these steps:
http://cksource.com/forums/viewtopic.php?f=5&t=15750
It says:
/ckeditor/plugins/link/dialogs/link.js
Changes

label:a.lang.common.target,'default':'notSet' --> label:a.lang.common.target,'default':'_blank'

label:a.lang.link.targetFrameName,'default':'' --> label:a.lang.link.targetFrameName,'default':'_blank'

It is also worth mentioning that if you include the ckeditor.js file twice in the same page, sometimes it causes trouble, thus detection in the CKEditor.tag files.