LDEV-3332: fix spring contexts declaration to find spring-beans-4.0.xsd locally at classpath.
Having spring module defined as a set of spring jars (as we do have in Wildfly), it is impossible for spring to resolve xsd schema mapping file 'spring.schemas' correctly. With set of jars, we have 5 instances of this file (with different content) under the same path 'META-INF/spring.schemas'. This ended up in spring not beeing able to map spring-beans-4.0.xsd URL path to local resource location. Finally, this xsd was loaded from internet - and that significantly slowed Wildfly startup.
With this fix LAMS is way (depends on internet connection) faster to start.