LDEV-3515: When creating a lesson do not preserve LD images' timestamps after all. Delete images after editing a LD in Flash Authoring.
Timestamps should not be preserved because if they do, Lesson LD's modification date is later than images' modification date. This triggers re-creation in Batik, which is not suitable for FLA LDs.
Now editing a LD in Flash Authoring deletes its images. They are stale anyway and would be deleted and re-created on the next access. But the risk of copying a stale image is not present anymore.