LDEV-4381 Removed requestSrc parameter
This parameter was responsible for passing integration server name to
LAMS which was used on closing authoring and refreshing integration
server page. It stopped working when authoring became Flash-less (as
there is not more "Close" button). And now we decided to finally remove
all its leftovers. This change also involves removal of
"isPostMessageToParent" parameter.