LDEV-4381 Remove obsolete lamsAltServerUrl setting
This setting was used by lams_ld_template and
lams_ld_template_blackboard projects. It was intended to solve the
original issue related to LAMS having to run as http:// for the video
streaming tool, but needing https:// to keep the iframe communication
happy. But once everything was switched to https there is no need in
this setting.