LDEV-4293 Change place where session gets mapped to login
UniversalLoginModule can be accessed without going through SsoHandler
when a session gets replicated. It means that it misses login ->
sessionID mapping which is needed for double login detection. The code
was rewritten to make sure login is properly mapped. It may be optimised
later so the mapping does not happen on every session start.