LDEV-4443: warning when editing activity that has been already used.
Currently when editing an assessment or voting activity in monitor, if an answer is changed, the data associated with that question is removed.

So if students have already completed the activity, their answers/marks will be removed and will shown as not answered.

We currently give no warning to monitors when they do this, so we should.