LDEV-4455 Move burning question to the learning.jsp
We need a change in the flow for Burning questions in learner, so the
burning question text for each question is entered as the students
answer select the answers. 

So each question would have a "burning question?" label that would
expand/collapse on request for students to enter the text as they answer
each question.