LKC-89 Move TBL Monitoring interface from lams-private
Add a switch that would display TBL UI monitoring interface. It
should be available only to the TBL template lessons.

Besides, modify the TBL UI monitor as in some cases the iRA (the first
assessment is not only just MCQ, but Assessment). So the change would be
to add as iRA the first assessment activity whether this is MCQ or
assessment.