LDEV-4540 Introduce hooks in command websockets
When a command is added, the JSON can contain "hookTrigger" and
"hookParams" attributes.
hookTrigger is just a string which uniquely identifies a page. If this
page set the same commandWebsocketHookTrigger, then command is consumed
and function defined on this page as commandWebsocketHook is run with
parameters taken from hookParams from JSON.
This way back end can tell exact page on front end to run any code it
defines itself.