LDEV-4540 Allow non-leaders see results page together with leader
Previously non-leader could see the results page only when the
leader goes to the next activity. If the leader is still on the
results page, a non-leader still sees the read-only questions page.
This way working with answer disclosing was pretty much
impossible so we allowed non-leaders see results page
simultaneously with leader.