LDEV-4054: Updating and fixing labels
There's quite a lot of work required with labels and it's not jsut translations but rearringing them according to language, string lenght, etc. Also we shouild be more accurate in naming users. When we are referring to students, we should call them students instead of users. Quite a few of these instances in this commit