LDEV-4605 Codify country for users
Change the country field to ISO_3166-1 alpha-2 codes (two letters).
So in when editing the profile, a user should be able to set his/her
country based on the list of countries in the ISO list.

Country should be a compulsory when editing profile, editing user
as group manager and in signup pages.

For integrations, if the country is not passed, then we should default
to a default country server.