LDEV-4440 Switch controllers to use @Autowired service beans
Also in cases when service bean in Controller has the same name as in
applicationContext, @Qualifier(..) can be removed. Which was done almost
in all cases, except for messageService due to "messageService" name
seems to be more preferable than adminMessageService,
centralMessageService and so on. Thus they were kept as before:
    @Autowired
    @Qualifier("centralMessageService")
    private MessageService messageService;