LDEV-4691 Reset a classloader when refreshing configuration cache
In WildFly 14 the Configuration bean can get created by one of LAMS EAR
modules, for example Image Gallery. This module's classloader does not
have access to java.sql.Connection. It should ask its parent for this
class, but context class loaders are not obliged to do it and apparently
WildFly module classloaders do not do it. We need to manually reset a
classloader for the given thread.