LDEV-4293 Prevent session invalidation exceptions from propagating
It seems that WildFly distributed session management may have bugs. On
session invalidation there is an attempt to close a websocket channel,
even if it does not exist. If it happens, an user gets in an endless
loop of accessing error.jsp page, trying to get session, trying to
invalidate his previous session, getting an error etc. We should at
least prevent the user from getting a visible error and let him try to
create a new session.