LDEV-5096 Move "Close" button to gate status page instead of monitoring
Allow closing all types of gates.
If a schedule gate gets opened, we do not clear its schedule anymore.
The user may have made a mistake when opening the gate and will
immediately close it again. We want the old schedule to remain.
When scheduled open gate job fires, it now checks if the gate is
actually closed. If it is already open, it does nothing.