LDEV-3387 Filter textarea data before (auto)save answers
Until now lams:textarea JS function was run only on textarea "change"
event. This event is run when the input looses focus. If an user keeps
writing in the same textare and autosave or time expired autosubmit
kicks in, data in the hidden input is not refreshed. Old data is sent
and stored in DB.
Now the data is refreshed before autosave and submit.