LDEV-5152 Use correct data for charts and item analysis.
For mark charts and stats we submitted data. It corresponds to
teachers' desire to see stats as soon as it is available, even when
learners have not finished an activity yet.
For item analysis we have to use only data from
learners who finished (not only submitted) the activity. It is because
lesson marks are updated only on activity finish and we need lesson
marks to calculate item analysis factors. An explanation for teachers
has been added.
If we want it to work the same as mark charts, we need to have lesson
marks to be updated on activity submit.