LDEV-5192 Open activity authoring initially maximised
If we set the now wider activity authoring pop up as the default size,
maximise button does not make sense, it will do nothing. So instead we
leave the standard pop up size but open it in maximised state. This way
teachers can return to standard size if they want to take a look at
authoring canvase in the background.