LDEV-5204 Initiate canvas copy on tool content copy on back-end
In previous version canvas was copied when an user opened it. Sometimes
in LAMS we copy content without opening a canvas. For example when using
Save As or Copy buttons in Authoring. We had to make sure that canvas is
always either copied on back-end or it gets opened and copied via
front-end.