LDEV-5204 Use separate file for each saved Whiteboard canvas
In original version all canvases were stored in a single JSON file. The
file would grow indefinitely. It would have to be stored in memory
whole. If it got corrupted, all canvases' data would be lost.
For all these reason now we store each canvas in its own file in
savedBoards/<whiteboard ID>.json
Each file is loaded only when corresponding canvas is used.
We can easily later do clean up - delete files which have not be used
for a year or so.
When doing a LAMS back up, the whole /savedBoards directory needs to be
backe up.