LDEV-5204 Embed images when exporting Whiteboard canvas
In vanilla Whiteboard images are referenced by absolute URLs. If we
upload content to another LAMS instance with a different URL, images
will not be found.
Now when exporting a LAMS LD we embed Whiteboard image base64 data
directly in tool exported XML. When we import the LD, we check if image
already exists on the server. If it does not, we recreate it from the
exported data.