LDEV-5204 Make Whiteboard always behave like it is in production mode
There is a problem with image loading in Whiteboard. Uploaded images get
stuck while being loaded in a copied canvas. For example after a learner
enters a grouped Whiteboard activity and a canvas gets cloned for his
session.
It does not occur when Whiteboard is in development mode. Maybe it has
to do with request proxying.
We now always run Whiteboard in development mode, but make it behave
like it is in production mode.