LDEV-3115 Properly set number of groups when opening a Run Sequence
In original LD a number of groups can be set in a grouping activity.
This number of groups can later change if a Monitor assigns learners to
groups in Monitoring UI.
When opening a Run Sequence, the grouping still holds old, lower value,
but there are more existing groups. This leads on validation errors on
re-save.
Now the number of groups is maximum of the original number of groups and
number of existing groups.