LDEV-5200 Adjusting icons colours and svg optimisation
I think this does it. However, there might be some issues with the embedding of SVG's css. As we compressed the files, the classes are called "A, B, C" and when different SVGs have the same class name, then the conflicts start to show.

Went thru all the tools but it's looking good now. There might be some others with system tools