LDEV-3767 Fix exception when user goes straight to finish page
If user is revisiting a completed Peer Review and both options for
showing results are turned off and also reflection is off, he gets
redirected straight to finish session action which requires a session
map ID parameter. This parameter is only available if finish session
action is called from a JSP page.