LDEV-4411 Stop using collection cache for now
Cache annotations were commented out so we can come back to them easily
in the future.
Collection cache is not detecting changes made to entities. When a new
entity is added, collection cache does not see the change.
Properly tweaking concurrency strategy does not seem possible at the
moment, so we just turn it off.