LDEV-3656 Clean up invalidated session from internal tracking mapping
These session should have been removed from the mapping when
SessionListener is called. Sometimes the listener is not called due to
WildFly internal bugs. We are left with a long list of stale sessions.
We should periodically clean up the mapping. Making it happen on runtime
stats fetch is as place as any other, plus it happens periodically.