LDEV-4293 Skip session creation when displaying error page
If the error was HTTP session being stale or incorrectly invalidated, we
should not try to use the same session to display the error page. The
error page should not require a session at all.
Also if making a decision which session to invalidate in case of double
login, always invalidate the other session and not current one as
otherwise it can lead to a loop of self-invalidating sessions.