LDEV-4538 At gate display only learners who have not passed it
Till now if a gate was opened, learners passed it and it was closed
again, it would show all those learners as "knocking at gate". It is
because they "attempted or completed" the activity and gate is closed.
Now TBL monitoring shows only learners who attempted the gate activity
but not completed it.