LDEV-5219 Change gate check flow in Peer Review
Criteria like mark hedging need to be validated before going to next
step, which can be finishing the activity. Current "check gate" JS code
does not allow it. Once the finish button is clicked, a check is made
and learner gets pushed to the next activity and whether his/her results
were saved or not.
Now we display results screen even if "show own ratings" or "show
ratings given to other learners" options are off. This allows to do
validation and store last criterion in DB. It also allows to show redo
questions button if this feature is on.
If there is no summary nor redo button, the results page automatically
attempts to go to next activity. If it succeeds, learner does not even
notice the blank results screen. If after Peer Activity there is a
closed gate, the learner will see single button and a message.